+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added Inventory for finished Compounds
 - Added document size and created date in fotmula attachements page
 - Fix search when replacing a material
+- PV Scale integration - WIP
 
 ### Version 9.9
 - Record replaced ingredients when finalising a formula in the generated PDF